Rust 2018, stable TryFrom
