+++ tests/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
-name = "stable_tests"
+name = "tests"
 version = "0.1.0"
 authors = ["Daniel Wagner-Hall <dawagner@gmail.com>"]
+edition = "2018"
 publish = false
 
 [dependencies]
